rename proto_helpers to just helpers
