+++ include/lbann/proto/CMakeLists.txt
@@ -2,6 +2,7 @@
 set_full_path(THIS_DIR_HEADERS
   init_image_data_readers.hpp
   proto_common.hpp
+  helpers.hpp
   )
 
 # Propagate the files up the tree
